Bencher tweaks
