+++ scripts/perf_summary.py
@@ -1,258 +1,69 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
 
+import argparse
+import json
 import os
 import sys
-import json
-import argparse
-import html
-import statistics
-from typing import List, Optional, Tuple
+from typing import List
 
-# Metric groups to chart over time. A chart is produced for every benchmark that
-# reports each metric.
-METRIC_GROUPS = [
-    ("throughput", "Throughput", "tx/s"),
-    ("latency", "Latency", "ms"),
-    ("memory", "Memory", "bytes"),
-    ("rate", "Rate", "ops/s"),
-]
-CHART_MAX_POINTS = 30
-CHART_COLUMNS = 4
-CHART_CELL_WIDTH = f"{100 // CHART_COLUMNS}%"
-EWMA_ALPHA = 0.3
-DEFAULT_REPOSITORY = "microsoft/CCF"
-METADATA_KEY = "__metadata"
+from perf_report import CHART_MAX_POINTS, METRIC_GROUPS
+from perf_report import METADATA_KEY
+from perf_report import PerfRun
+from perf_report import commit_url, run_url
+from perf_report import benchmarks_with_metric, jobid_sort_key
+from perf_report import list_perf_files, load_bencher_file, load_perf_data
+from perf_report import render_metric_group, render_runs_table
+from perf_report import render_perf_summary
 
-PerfRun = Tuple[str, Optional[str], Optional[str], dict]
-ChartSeries = List[Tuple[str, float]]
+MAIN_HISTORY_POINTS = 10
 
 
-def jobid_sort_key(name: str) -> Tuple[int, object]:
-    """Order perf files chronologically by their numeric job id.
-
-    File names have the form ``<run_id>-<run_number>-<run_attempt>.json`` where
-    each component increases over time, so ordering by the integer components
-    gives chronological order. Falls back to the name for unexpected formats.
-    """
+def comparison_sort_key(name: str) -> tuple:
     stem = name[:-5] if name.endswith(".json") else name
-    try:
-        return (0, tuple(int(part) for part in stem.split("-")))
-    except ValueError:
-        return (1, name)
+    index = stem.rsplit("-", 1)[-1]
+    return (0, int(index)) if index.isdigit() else jobid_sort_key(name)
 
 
-def list_perf_files(directory: str) -> List[str]:
-    """Return perf files in the directory, ordered chronologically (oldest first)."""
-    if not os.path.isdir(directory):
-        return []
-    files = [
-        name
-        for name in os.listdir(directory)
-        if os.path.isfile(os.path.join(directory, name))
-    ]
-    return sorted(files, key=jobid_sort_key)
-
-
-def run_label(name: str) -> str:
-    """Short x-axis label for a perf file: the run number when available."""
-    stem = name[:-5] if name.endswith(".json") else name
-    parts = stem.split("-")
-    return parts[1] if len(parts) >= 2 else stem
-
-
-def run_url(name: str) -> Optional[str]:
-    """GitHub Actions URL for a perf file, when the run id can be parsed."""
-    stem = name[:-5] if name.endswith(".json") else name
-    parts = stem.split("-")
-    if not parts or not parts[0].isdigit():
-        return None
-
-    server_url = os.environ.get("GITHUB_SERVER_URL", "https://github.com").rstrip("/")
-    repository = os.environ.get("GITHUB_REPOSITORY", DEFAULT_REPOSITORY)
-    return f"{server_url}/{repository}/actions/runs/{parts[0]}"
-
-
-def commit_url(metadata: dict) -> Optional[str]:
-    """GitHub commit URL from perf metadata, when available."""
-    commit = metadata.get("commit")
-    if not isinstance(commit, str) or not commit:
-        return None
-
-    server_url = metadata.get("server_url") or os.environ.get(
-        "GITHUB_SERVER_URL", "https://github.com"
-    )
-    repository = metadata.get("repository") or os.environ.get(
-        "GITHUB_REPOSITORY", DEFAULT_REPOSITORY
-    )
-    if not isinstance(server_url, str) or not isinstance(repository, str):
-        return None
-    return f"{server_url.rstrip('/')}/{repository}/commit/{commit}"
-
-
-def load_perf_data(directory: str, files: List[str]) -> List[PerfRun]:
-    """Load (label, run_url, commit_url, data) for each readable perf file."""
-    loaded: List[PerfRun] = []
-    for name in files:
-        try:
-            with open(os.path.join(directory, name), "r") as f:
-                data = json.load(f)
-        except (OSError, json.JSONDecodeError):
-            continue
-        if isinstance(data, dict):
-            metadata = data.get(METADATA_KEY, {})
-            if not isinstance(metadata, dict):
-                metadata = {}
-            loaded.append((run_label(name), run_url(name), commit_url(metadata), data))
-    return loaded
-
-
-def metric_value(data: dict, benchmark: str, metric: str) -> Optional[float]:
-    """Return the numeric value of a benchmark metric, or None if absent."""
-    metrics = data.get(benchmark)
-    if not isinstance(metrics, dict):
-        return None
-    entry = metrics.get(metric)
-    if not isinstance(entry, dict):
-        return None
-    value = entry.get("value")
-    return value if isinstance(value, (int, float)) else None
-
-
-def benchmarks_with_metric(loaded: List[PerfRun], metric: str) -> List[str]:
-    """Sorted names of benchmarks that report the given metric in any run."""
-    names = set()
-    for _, _, _, data in loaded:
-        for benchmark in data:
-            if benchmark == METADATA_KEY:
-                continue
-            if metric_value(data, benchmark, metric) is not None:
-                names.add(benchmark)
-    return sorted(names)
-
-
-def ewma(values: List[float], alpha: float = EWMA_ALPHA) -> float:
-    """Return the exponentially weighted moving average of the values."""
-    average = values[0]
-    for value in values[1:]:
-        average = alpha * value + (1 - alpha) * average
-    return average
-
-
-def repeated_values(value: float, count: int) -> str:
-    """Render a constant series for every chart category."""
-    return ", ".join(f"{value:.2f}" for _ in range(count))
-
-
-def render_mermaid_xychart(
-    series: ChartSeries,
-    benchmark: str,
-    metric: str,
-    unit: str,
-) -> str:
-    """Render a Mermaid xychart line chart for a single benchmark metric."""
-    ordered_series = list(reversed(series))
-    labels = ", ".join(f'"{label}"' for label, _ in ordered_series)
-    raw_values = [value for _, value in ordered_series]
-    values = ", ".join(f"{value:.2f}" for value in raw_values)
-    chronological_values = [value for _, value in series]
-    baseline = ewma(chronological_values)
-    sigma = (
-        statistics.pstdev(chronological_values) if len(chronological_values) > 1 else 0
-    )
-    lines = [
-        f"<h4>{html.escape(benchmark)}</h4>",
-        "",
-        "```mermaid",
-        "---",
-        "config:",
-        "    xyChart:",
-        "        width: 220",
-        "        height: 320",
-        "        showTitle: false",
-        "        xAxis:",
-        "            labelFontSize: 10",
-        "            titleFontSize: 12",
-        "        yAxis:",
-        "            labelFontSize: 8",
-        "            titleFontSize: 12",
-        "            showTitle: false",
-        "    themeVariables:",
-        "        xyChart:",
-        '            plotColorPalette: "#003E7E, #62B5E5, #C7E9FB, #C7E9FB"',
-        "---",
-        "xychart horizontal",
-        f"    x-axis [{labels}]",
-        f'    y-axis "{metric} ({unit})"',
-        f"    line [{values}]",
-        f"    line [{repeated_values(baseline, len(raw_values))}]",
-        f"    line [{repeated_values(baseline - sigma, len(raw_values))}]",
-        f"    line [{repeated_values(baseline + sigma, len(raw_values))}]",
-        "```",
-        "",
-    ]
-    return "\n".join(lines)
-
-
-def render_chart_table(
-    loaded: List[PerfRun], benchmarks: List[str], metric: str, unit: str
-) -> str:
-    """Render benchmark charts in a four-column table."""
-    lines = ['<table width="100%">']
-    for index, benchmark in enumerate(benchmarks):
-        if index % CHART_COLUMNS == 0:
-            lines.append("<tr>")
-        lines.append(f'<td valign="top" width="{CHART_CELL_WIDTH}">')
-        series = [
-            (label, value)
-            for label, _, _, data in loaded
-            if (value := metric_value(data, benchmark, metric)) is not None
+def list_comparison_files(path: str) -> List[str]:
+    if os.path.isdir(path):
+        files = [
+            os.path.join(path, name)
+            for name in sorted(os.listdir(path), key=comparison_sort_key)
+            if name.endswith(".json") and os.path.isfile(os.path.join(path, name))
         ]
-        lines.append(render_mermaid_xychart(series, benchmark, metric, unit))
-        lines.append("</td>")
-        if index % CHART_COLUMNS == CHART_COLUMNS - 1:
-            lines.append("</tr>")
-    remaining = len(benchmarks) % CHART_COLUMNS
-    if remaining:
-        for _ in range(CHART_COLUMNS - remaining):
-            lines.append(f'<td valign="top" width="{CHART_CELL_WIDTH}"></td>')
-        lines.append("</tr>")
-    lines.append("</table>")
-    lines.append("")
-    return "\n".join(lines)
+        if files:
+            return files
+        raise FileNotFoundError(f"No JSON files found in {path}")
+    return [path]
 
 
-def render_runs_table(loaded: List[PerfRun]) -> str:
-    """Render a compact table of run labels, Actions runs, and commits."""
-    lines = ["### Runs", "", "| Run | Actions | Commit |", "| --- | --- | --- |"]
-    for label, run, commit, data in reversed(loaded):
+def load_comparison_data(path: str, label: str) -> List[PerfRun]:
+    files = list_comparison_files(path)
+    count = len(files)
+    runs = []
+    for index, file_path in enumerate(files, 1):
+        data = load_bencher_file(file_path)
         metadata = data.get(METADATA_KEY, {})
-        commit_sha = metadata.get("commit") if isinstance(metadata, dict) else None
-        short_commit = commit_sha[:8] if isinstance(commit_sha, str) else ""
-        run_link = f"[run]({run})" if run else ""
-        commit_link = f"[{short_commit}]({commit})" if commit and short_commit else ""
-        lines.append(f"| {label} | {run_link} | {commit_link} |")
-    lines.append("")
-    return "\n".join(lines)
+        if not isinstance(metadata, dict):
+            metadata = {}
+        run_id = metadata.get("run_id")
+        run_label = label if count == 1 else f"{label} {index}"
+        run = run_url(f"{run_id}.json") if isinstance(run_id, str) else None
+        runs.append((run_label, run, commit_url(metadata), data))
+    return runs
 
 
-def render_metric_group(
-    loaded: List[PerfRun], metric: str, title: str, unit: str
-) -> str:
-    """Render one chart per benchmark that reports the given metric."""
-    benchmarks = benchmarks_with_metric(loaded, metric)
-    lines = [f"## {title} ({unit})", ""]
-    if not benchmarks:
-        lines.append(f"_No benchmarks with a `{metric}` metric found._")
-        lines.append("")
-        return "\n".join(lines)
+def render_comparison(main_runs: List[PerfRun], comparison_runs: List[PerfRun]) -> str:
+    main_history = main_runs[-MAIN_HISTORY_POINTS:]
+    loaded = [*main_history, *comparison_runs]
+    comparison_benchmarks = {
+        metric: benchmarks_with_metric(comparison_runs, metric)
+        for metric, _, _ in METRIC_GROUPS
+    }
+    if not any(comparison_benchmarks.values()):
+        raise ValueError("No supported metrics found in comparison results")
 
-    lines.append(render_chart_table(loaded, benchmarks, metric, unit))
-    return "\n".join(lines)
-
-
-def render_perf_summary(loaded: List[PerfRun]) -> str:
-    """Render all perf metric groups as markdown."""
     lines = [
         "# Performance summary",
         "",
@@ -261,7 +72,19 @@ def render_perf_summary(loaded: List[PerfRun]) -> str:
         render_runs_table(loaded),
     ]
     for metric, title, unit in METRIC_GROUPS:
-        lines.append(render_metric_group(loaded, metric, title, unit))
+        benchmarks = comparison_benchmarks[metric]
+        if not benchmarks:
+            continue
+        lines.append(
+            render_metric_group(
+                loaded,
+                metric,
+                title,
+                unit,
+                benchmarks=benchmarks,
+                reference_loaded=main_history if main_history else None,
+            )
+        )
     return "\n".join(lines)
 
 
@@ -275,13 +98,29 @@ def main() -> None:
         default="perf",
         help="Directory containing the perf data files (default: perf)",
     )
+    parser.add_argument(
+        "--compare",
+        help="PR bencher JSON file or directory to append after main history",
+    )
+    parser.add_argument(
+        "--label",
+        default="PR",
+        help="Label for comparison results (default: PR)",
+    )
     args = parser.parse_args()
 
     files = list_perf_files(args.directory)
-
     recent = files[-CHART_MAX_POINTS:]
-    loaded = load_perf_data(args.directory, recent)
-    print(render_perf_summary(loaded))
+    main_runs = load_perf_data(args.directory, recent)
+    if args.compare:
+        try:
+            comparison_runs = load_comparison_data(args.compare, args.label)
+        except (OSError, json.JSONDecodeError, ValueError) as exc:
+            print(f"Error: {exc}", file=sys.stderr)
+            sys.exit(1)
+        print(render_comparison(main_runs, comparison_runs))
+    else:
+        print(render_perf_summary(main_runs))
 
 
 if __name__ == "__main__":
